# backend/docs/swagger.yaml
basePath: /
definitions:
...
  datauptwo_app_dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:18421
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: datauptwo 数据管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: datauptwo API
  version: "1.0"
paths:
...
  /api/v1/sync/stations:
    get:
      summary: 获取站点列表
      tags: [Sync]
      parameters:
        - in: query
          name: page
          type: integer
        - in: query
          name: pageSize
          type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
    post:
      summary: 创建站点
      tags: [Sync]
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              name:
                type: string
              code:
                type: string
              url:
                type: string
              status:
                type: string
              description:
                type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
  /api/v1/sync/stations/{id}:
    put:
      summary: 更新站点
      tags: [Sync]
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: body
          schema:
            type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
    delete:
      summary: 删除站点
      tags: [Sync]
      parameters:
        - in: path
          name: id
          type: integer
          required: true
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
  /api/v1/sync/history:
    get:
      summary: 获取同步历史
      tags: [Sync]
      parameters:
        - in: query
          name: page
          type: integer
        - in: query
          name: pageSize
          type: integer
        - in: query
          name: stationId
          type: integer
        - in: query
          name: startDate
          type: string
        - in: query
          name: endDate
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
...
  /api/v1/sync/register:
    post:
      summary: Agent 自动注册
      tags: [Sync]
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              stationId:
                type: string
              stationName:
                type: string
              url:
                type: string
              password:
                type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
...
swagger: "2.0"